solved fr1-S

+++ Excited/fr1-S.cpp
@@ -1,7 +1,7 @@
 #include <cstdio>
 
 int main() {
-    int ways[60]; ways[0] = 0; ways[1] = 1;
+    long long ways[60]; ways[0] = 0; ways[1] = 1;
     for(int i = 2; i < 60; i++) {
         ways[i] = ways[i - 1] + ways[i - 2];
     }
@@ -9,7 +9,7 @@ int main() {
     std::scanf("%d", &n);
     while(n--) {
         std::scanf("%d %d", &a, &b);
-        std::printf("%d\n", ways[b - a + 1]);
+        std::printf("%lld\n", ways[b - a + 1]);
     }
     return 0;
 }
